+++ src/main/webapp/js/paginas/secoes/conta/secaoDadosPessoais/modalAlterarDadosPessoais.js
@@ -1,7 +1,7 @@
 import { montarClientePorForm } from "../../../../script.js";
 import { atualizarCliente } from "@api/cliente.api.js";
 import { FormularioDadosPessoais } from "@componentes/forms/formDadosPessoais.js";
-import { Modal } from "@componentes/modal.js";
+import { Modal } from "@componentes/common/modal.js";
 import { alertarErro } from "@api/alertErro.js";
 import { localStorageKeys } from "../../../../localStorage.js";
